Introduce a client for Logit's OpenSearch API
As mentioned when installing RestClient, I did try to use the official
OpenSearch Ruby client gem but it ultimately didn't seem to be a good
fit for our limited use case.

I've opted for basic auth API access for now purely because that's the
access that I was able to get for myself. We might decide to switch on
API key access and have to update how this client works before we're
ready to start using this in production.