Introduce the publishing latency SLI
This class will represent our (for now, one-and-only) SLI. It's going to
fetch data from Logit, turn those into metrics and push them to
Prometheus for use in Grafana.

I'm mostly introducing this now in this bare state to confirm that we
can find classes and run specs.

The spec example's name describes how I'm thinking of building out this
class. Since its configuration will be provided by environment
variables, if it accesses those internally and can avoid needing
constructor and method parameters, there are less places to update when
changes are needed.