Introduce ContentStoreEvents
Similarly to WhitehallEvents, but fetches the corresponding publish
event data from Content Store's logs using the `govuk_request_id`s of
the Whitehall ones.

We also don't restrict the `to_time` like we do with WhitehallEvents.
That's intentional since, as explained in the commit that introduced
WhitehallEvents, Whitehall log events are our "start" events and these
Content Store log events are our "completion" events. We're avoiding
retrieving up-to-the-minute start events in order to give the end-to-end
process time to complete, so once we've got some start events, we want
to get our hands on any and all matching completion events, from as
early as the beginning of our time window and upto as recent as right
now.

The `app_name` supplied here is a temporary one while Content Store is
still in the process of being migrated to PostgreSQL. We expect this
value to change twice in the coming weeks, first to
"content-store-postgresql-branch", around the time that the PostgreSQL
instance becomes the Content Store Proxy's primary upstream (and ideally
before "content-store-mongo-main" becomes unavailable) and then later to
just "content-store" once Content Store Proxy vacates the app name that
it's been squatting.