Github workflow for deploying to integration
These workflow setups seem pretty much identical across projects. So, as
standard, this:
1. builds the container image
2. pushes it to ECR
3. updates the image tag in govuk-helm-charts

The cronjob (that we're planning to add to govuk-helm-charts) will then
pull the released image the next time it runs.

Note: unlike most of our projects, this one will only be able to be
deployed to a single, hard-coded target environment.

Until we're ready to put this into production, we'll test this by
deploying to integration. After that, we'll need to change this
environment setting.

As things stand (i.e. only having the one Publishing Latency SLI right
now), once we've switched to production, I don't think there's much
value in allowing deployments to additional environments even for
testing purposes, because:
1. as far as I can tell, only production ever has enough publish events
   in it to provide this SLI with any data to work with
2. we'd either have to continuously run the cron job in those other
   environments so that they're readily available on the off-chance that
   we'll want to use them for testing once in a while or we'd have to
   get configuration changes merged to govuk-helm-charts before and
   after any testing session