Introduce Error singleton
For sending exceptions to Sentry. One method sends and silences the
exception, the other sends and re-raises it.

`Sentry.init` needed to be called before we used any Sentry methods, so
while it's a little bit gross, this ended up feeling like the right
place.

Our Helm chart already includes the SENTRY_DSN, SENTRY_CURRENT_ENV and
SENTRY_RELEASE in our job's environment, by the way, so Sentry should be
configured out-of-the-box. (Othewise, we'd supply configuration to
`Sentry.init`.)