Fix reference to tag name in deploy workflow
Copied from Sean's fixed merged into other GOV.UK repositories:

There is a known bug in GitHub Actions that `github.ref_name` to not be
populated when the workflow was triggered due to a release tag. This
uses an alternative field to reference the release tag value. This fixes
intermittent failures in running the deploy workflow.