Install deployment target environment input
We currently only allow deploying to a single environment and we don't
expect that to change very often or permanently.

In the interests of ease of future development, I think it makes sense
to leave this in a state that's ready to add options to -- thinking of
testing changes in integration (note: that would also require enabling
the CronJob for that environment in the Helm charts repository).