Turn off automatic deployments
Automatic deployments are designed with integration in mind and don't
work when production is the target (as is the case for us).

This project requires manual deployment. (And already did require that
before this commit, only it didn't know it yet.)

I've made the bare minimum change to turn this feature off: removing the
`release` trigger from the workflow file. But I have left some of its
supporting code in place. Github workflows setups are mostly identical
across our code bases and occasionally multiple projects receive fixes
and updates to these files simultaneously. I'm optimistically embracing
the idea that the closer we can keep this project's versions of these
files, the more likely it is that we'll receive fixes and updates when
those other projects do.